+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [feat: (#43) bitwise OR operator (`|`)](https://github.com/amp-lang/amp/issues/43)
 - [feat: (#44) `!` logical NOT operator](https://github.com/amp-lang/amp/issues/44)
 - [feat: (#44) `&&` logical AND operator](https://github.com/amp-lang/amp/issues/44)
+- [feat: (#44) `||` logical OR operator](https://github.com/amp-lang/amp/issues/44)
 
 ### Library
 - feat: implement standard library, I guess